requirements.txt
python-telegram-bot==21.11
python-socketio[aiohttp]==5.11.1
aiohttp==3.9.3
pyyaml==6.0.1
telethon==1.32.1
cryptg==0.4.0  # For faster encryption/decryption in Telethon
pillow==10.2.0  # For handling images in Telethon
pytest>=7.4.0
pytest-asyncio>=0.23.0
